refactor(scheduler): remove directly sending whole states
