+++ Grains/usage-tracker/ISchedulerGrain.cs
@@ -8,12 +8,12 @@ public interface ISchedulerGrain : ISchrodingerGrain, Orleans.IGrainWithStringKe
     Task<IReadOnlyDictionary<string, RequestAccountUsageInfo>> GetFailedImageGenerationRequestsAsync();
     Task<IReadOnlyDictionary<string, RequestAccountUsageInfo>> GetStartedImageGenerationRequestsAsync();
     Task<IReadOnlyDictionary<string, RequestAccountUsageInfo>> GetPendingImageGenerationRequestsAsync();
-    Task<IReadOnlyDictionary<string, BlockedRequestInfo>> GetBlockedImageGenerationRequestsAsync();
+    Task<IEnumerable<BlockedRequestInfoDto>> GetBlockedImageGenerationRequestsAsync();
     Task AddImageGenerationRequest(string requestId, string childId, long requestTimestamp);
     Task<List<ApiKey>> AddApiKeys(List<ApiKeyEntryDto> apiKeyEntries);
     Task<List<ApiKey>> RemoveApiKeys(List<ApiKey> apiKeys);
     Task<IReadOnlyList<ApiKeyEntryDto>> GetAllApiKeys();
-    Task<SchedulerState> GetImageGenerationStates();
+    Task<Dictionary<string, IEnumerable<RequestAccountUsageInfoDto>>> GetImageGenerationStates();
     Task<Dictionary<string, ApiKeyUsageInfo>> GetApiKeysUsageInfo();
     Task<bool> IsOverloaded();
     Task FlushAsync();
