+++ Shared/SchedulerAPISchema.cs
@@ -30,18 +30,6 @@ public class RemoveApiKeyResponseFailed(string error) : RemoveApiKeyAPIResponse
     public string Error { get; set; } = error;
 }
 
-public class RequestAccountUsageInfoDto
-{
-    public string RequestId { get; set; } = "";
-    public string RequestTimestamp { get; set; } = "";
-    public string StartedTimestamp { get; set; } = "";
-    public string FailedTimestamp { get; set; } = "";
-    public string CompletedTimestamp { get; set; } = "";
-    public int Attempts { get; set; } = 0;
-    public ApiKey? ApiKey { get; set; } = null;
-    public string ChildId { get; set; } = "";
-}
-
 public abstract class ImageGenerationStatesResponse { }
 
 public class ImageGenerationStatesResponseOk<T>(T imageGenerationStates) : ImageGenerationStatesResponse
@@ -104,12 +92,6 @@ public class ForceRequestExecutionResponseFailed(string error) : ForceRequestExe
 {
     [JsonPropertyName("error")]
     public string Error { get; set; } = error;
-}
-
-public class BlockedRequestInfoDto
-{
-    public string? BlockedReason { get; set; } = "";
-    public RequestAccountUsageInfoDto RequestInfo { get; set; }
 }
 
 public abstract class BlockedRequestResponse { }
